+++ authing/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.0.4
+
+* add deleteAccount API
+* add login by scanning QR code API
+* after register, return id token
+
+## 1.0.3
+
+* update doc
+
+## 1.0.2
+
+* update doc
+
 ## 1.0.1
 
 * update doc